% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_Bayes.R
\name{estimate_R_Bayes}
\alias{estimate_R_Bayes}
\title{Estimate effective reproduction number in a Bayesian framework}
\usage{
estimate_R_Bayes(
  donset,
  ddiag,
  local,
  t_from,
  SI_mean,
  SI_std,
  t_window = 7,
  t_breaks = NULL,
  maxDate = NULL,
  weights = NULL,
  withInfectTimes = TRUE,
  delayAdjust = TRUE,
  incub_mean = 5.1,
  incub_std = 3,
  date0 = as.Date("2020-1-31"),
  burn = 500,
  iter = 5000
)
}
\arguments{
\item{donset}{A vector of symptom onset dates (of class \code{Date}).}

\item{ddiag}{A vector of diagnosis dates (of class \code{Date}), 
of equal length to \code{donset}; alternatively \code{NULL} 
if working only with symptom onset dates, see 'Details' section.}

\item{local}{A logical vector, of equal length to \code{donset}
indicating whether the case is local or imported.}

\item{t_from}{Date (of class \code{Date}) from which to start estimating
 effective reproduction numbers (instantaneous or period-specific);
see 'Details' section.}

\item{SI_mean}{Mean for the serial interval distribution, either a single
value or two values defining a range; see 'Details' section.}

\item{SI_std}{Standard Deviation for the serial interval distribution, 
either a single value or two values defining a range; see 'Details' section.}

\item{t_window}{Window length for estimating a rolling instantaneous 
effective reproduction number, see 'Details' section.}

\item{t_breaks}{Dates defining specific time windows for estimating the 
effective reproduction number, or \code{NULL} for a rolling 
instantaneous reproduction number, see 'Details' section.}

\item{maxDate}{Maximum date for which to estimate the effective 
reproduction number; if \code{NULL}, the maximum diagnosis date or 
symptom onset date is used.}

\item{weights}{An optional vector of weights for the cases. If \code{NULL} 
(the default), all cases are given equal weights of 1.}

\item{withInfectTimes}{Extrapolate infection times and use those for 
estimating the effective reproduction number. If \code{TRUE} 
(the default), a Gamma distribution with mean \code{incub_mean} and 
standard deviation \code{incub_std} will be used to impute the
infection times.}

\item{delayAdjust}{Adjust for delays in case ascertainment, using the 
distribution of times from symptom onset to diagnosis; 
see 'Details' section}

\item{incub_mean}{Mean incubation period, used for extrapolating 
infection times from symptom onset dates.}

\item{incub_std}{Standard deviation of the incubation period distribution,
used for extrapolating infection times from symptom onset dates.}

\item{date0}{"Date zero" from which to start counting when converting the
provided Date vectors to integers for MCMC fitting. Also, imputation
of symptom onset and infection dates will start after this date.}

\item{burn}{Number of burn-in MCMC iterations}

\item{iter}{Number of sampling MCMC iterations}
}
\value{
An object of class 'bayEStim'. This is a list containing the following elements:
  \describe{
    \item{$model}{The fitted model; an object of class \code{rjags}}

    \item{$mcmcSamples}{An object of class \code{coda} containing the MCMC samples for
        the various model parameters.}

    \item{$config}{A list of all configuration parameters used in the model.}
  }

Objects of class 'bayEStim' can be \link[summary.bayEStim]{summarized} and 
\link[plot.bayEStim]{plotted}, see the respective methods for details.
}
\description{
\code{estimate_R_Bayes} estimates the effective reproduction number (R_e)
using the method of Cori et al, but in a Bayesian framework allowing 
(a) uncertainty in the serial interval distribution, 
(b) imputation of missing symptom onset dates, 
(c) extrapolation of infection dates, and (d) adjusting for 
delays between diagnosis, symptom onset and infection ("nowcasting")
}
\details{
\code{estimate_R_Bayes()} allows flexibility in how to estimate
    the effective reproduction number R_e, using different function arguments.

At the most basic, a vector of symptom onset dates must be provided 
(argument \code{donset}) and on this basis the R_e will be estimated, if 
\code{withInfectTimes = FALSE}. If \code{withInfectTimes = TRUE} (the 
default), the R_e is estimated on the basis of the infection times 
imputed using a Gamma distribution with mean \code{incub_mean} and 
standard deviation \code{incub_std}. 

If a vector of diagnosis dates is also provided (argument \code{ddiag}), 
then \code{donset} can have missing values, and these will be imputed 
using the distribution of the time between symptom onset and diagnosis
(assumed to be Gamma, with parameters estimated from the data). 
In addition, if diagnosis dates are provided, it is possible to use this
distribution (of the time between symptom onset and diagnosis) to 
"nowcast" the epidemic, if the argument \code{delayAdjust} is set to 
\code{TRUE}: the number of cases with symptom onset at each time is 
divided by the probability of ascertainment by date \code{maxDate}.

If argument \code{t_breaks} is \code{NULL} (the default), the function
estimates the instantaneous effective reproduction number, using rolling
weekly windows between dates \code{t_start} and \code{maxDate}. The size
of the window is adjusted by the argument \code{t_window}. 
Alternatively, \code{t_breaks} can contain a vector of dates that define
(together with \code{t_breaks} and \code{maxDate}) specific periods for
which the effective reproduction number will be estimated.

Arguments \code{SI_mean} and \code{SI_std} can receive either one or two
values. If one value is provided, the parameter is assumed fixed. 
If two values are provided, these define a range from which the parameter
is randomly sampled (using a uniform distribution). This allows 
to incorporate uncertainty in estimating the serial interval 
distribution parameters.
}
\examples{
# We'll eventually put some examples here

}
\references{
\itemize{
 \item Cori A, Ferguson NM, Fraser C, Cauchemez S. A new framework and 
 software to estimate time-varying reproduction numbers during epidemics. 
 \href{https://academic.oup.com/aje/article-lookup/doi/10.1093/aje/kwt133}{Am J Epidemiol} 
 2013;178(9):1505–12 (\href{https://www.ncbi.nlm.nih.gov/pubmed/24043437}{PubMed})
 
 \item Thompson RN, Stockwin JE, van Gaalen RD, et al. Improved inference of 
 time-varying reproduction numbers during infectious disease outbreaks. 
 \href{https://linkinghub.elsevier.com/retrieve/pii/S1755-4365(19)30035-0}{Epidemics} 
 2019;29:100356 (\href{https://www.ncbi.nlm.nih.gov/pubmed/31624039}{PubMed})
}
}
